Add enforcement that all tariff schedules should start at hour 0.
Previously, it was allowable to begin a tariff schedule not at 0, and
expect the schedule to wrap back to the previous rate. This added only
a slight convenience and invites bugs, so now all tariff schedules
should begin with 0 since it is assumed that the time is the hour when
that price takes effect.