Fix out of bounds error from empty schedule.
If an empty schedule is submitted by the scheduler right as the
pilot_signals and charging_rates matrices needed to be expanded, the next
iteration will yield an out of bounds error when EVSE pilots are updated.
This fix ensures the pilot_signals and charging_rates are at least
long enough such that the current iteration may be indexed into.