Add more accurate linear 2 stage battery fitting and charging models.
Fitting a battery to data such that tail behavior is maximized can be done
by binary searching the solution to the battery differential equation.
Likewise, charging a linear 2 stage battery can be done with higher precision
if the solution to the battery differential equation is used directly rather
than an approximation. Both of these changes are implemented.