Add version checking to serializer.
The new serializer checks the version of acnportal that
is loading the object against the version of acnportal
that dumped the object, and raises a warning if there is
a mismatch. This is also done for versions of the
dependencies numpy and pandas, as these may cause issues
if there is a version mismatch.