Fix test_cr_interface_func to pass with definition.
last_actual_charging_rate has a new definition that
doesn't return charging rates delivered to vehicles that
arrived in the last timestep, as the charging rate would
have been delivered to the EVSE before the new vehicle
was plugged in. The test still assumed the old
definition. Namely, at timestep 0, an EV plugs in, but
since there was no previous iteration, there shouldn't
be a last charging rate, so the function should return
{}.