Refactor EVSE classes to use a BaseEVSE class.
EVSEs have different salient parameters in their
constructors depending on their types. So that
constructors don't need to retain redundant arguments,
a BaseEVSE class is created from which all EVSE classes
inherit.

Also, a bug in which the atol argument was not used
for certain boundaries of the DeadbandEVSE's feasible
set was fixed, and tests were added for the atol
functionality.