Add schedule function to GymAlgorithm.
If an RL model is deployed in production (namely, to
ACN-Live), it must fit into the workflow of a
Simulator running, calling a scheduler, and receiving a
schedule. If this is to be done, GymAlgorithm needs to
actually implement a schedule function, which entails
running the model to return a schedule, and updating
the environment for use by the model.

Also, the step function of the simulator should only
be accessible in training, as in real life such a
function would not be realistic. A new interface class
was created to reflect this.