Added a gym training interface.
It's envisioned that the workflow of an RL researcher
using these tools would be to train a model by inputting
an environment to a model and calling model.learn, then
using said model as part of an algorithm in sim.run.
For the former, an env step function is required, but the
latter should not need it, especially if live deployment
is desired. So, the step function was assigned to a new
interface and checks were added to ensure the right
kind of interface was being used for the use case.