Add options for serializer to accept a filepath or buffer.
The serializer now accepts an argument specifying a filepath
or buffer to which the serialized object may be directly
written. If this argument is not provided, a JSON string is
returned.

The deserializer accepts a JSON string as before, or a filepath
or buffer from which the JSON string can be read.

Add tests for this functionality, and documentation.

Note: Some of this code was taken from pandas v1.0.1. See
source for attribution.